rename lynch to condemn
