+++ Games/types/Mafia/items/Suit.js
@@ -16,7 +16,7 @@ module.exports = class Suit extends Item {
       death: true,
       reveal: true,
       investigate: true,
-      lynch: true,
+      condemn: true,
     };
 
     super.hold(player);
