+++ Games/types/Mafia/roles/Village/Seer.js
@@ -4,12 +4,12 @@ module.exports = class Seer extends Role {
   constructor(player, data) {
     super("Seer", player, data);
     this.alignment = "Village";
-    this.cards = ["VillageCore", "WinWithVillage", "RevealScumToRole"];
+    this.cards = ["VillageCore", "WinWithVillage", "RevealEvilPlayersToSelf"];
 
     this.appearance = {
       self: "real",
       reveal: "real",
-      lynch: "real",
+      condemn: "real",
       death: "Villager",
       investigate: "real",
     };
