+++ Games/types/Mafia/roles/cards/WinWithCult.js
@@ -1,5 +1,5 @@
 const Card = require("../../Card");
-const { PRIORITY_WIN_CHECK_DEFAULT } = require("../../const/Priority");
+const { PRIORITY_WIN_CHECK_DEFAULT, PRIORITY_SUNSET_DEFAULT } = require("../../const/Priority");
 
 module.exports = class WinWithCult extends Card {
   constructor(role) {
@@ -44,13 +44,13 @@ module.exports = class WinWithCult extends Card {
           }
           let isOverthrow, target;
           for (const action of this.game.actions[0]) {
-            if (action.target && action.hasLabels(["lynch", "overthrow"])) {
+            if (action.target && action.hasLabels(["condemn", "overthrow"])) {
               isOverthrow = true;
               target = action.target;
             } else if (
               !isOverthrow &&
               action.target &&
-              action.hasLabel("lynch")
+              action.hasLabel("condemn")
             ) {
               target = action.target;
             }
@@ -59,14 +59,14 @@ module.exports = class WinWithCult extends Card {
         },
         action: {
           labels: ["kill"],
-          priority: PRIORITY_LYNCH_REVENGE,
+          priority: PRIORITY_SUNSET_DEFAULT,
           run: function () {
             if (this.target.role.name === "Seer") {
               if (!this.game.guessedSeers) {
                 this.game.guessedSeers = [];
               }
               this.game.guessedSeers.push(this.target);
-              this.target.kill("lynchRevenge", this.actor);
+              this.target.kill("condemnRevenge", this.actor);
             }
           },
         },
@@ -93,13 +93,13 @@ module.exports = class WinWithCult extends Card {
           }
           let isOverthrow, target;
           for (const action of this.game.actions[0]) {
-            if (action.target && action.hasLabels(["lynch", "overthrow"])) {
+            if (action.target && action.hasLabels(["condemn", "overthrow"])) {
               isOverthrow = true;
               target = action.target;
             } else if (
               !isOverthrow &&
               action.target &&
-              action.hasLabel("lynch")
+              action.hasLabel("condemn")
             ) {
               target = action.target;
             }
