+++ data/roles.js
@@ -785,7 +785,7 @@ const roleData = {
       alignment: "Village",
       description: [
         "Knows all of the Mafia and Cult at the start of the game.",
-        "When lynched, Mafia and Cult have a chance to guess who the Seer is.",
+        "When condemned, Mafia and Cult have a chance to guess who the Seer is.",
         "On a correct guess, the Seer dies and the Mafia or Cult wins.",
         "Appears as villager on death.",
       ],
